requirements_additions.txt
# Additional Requirements for Infrastructure Implementation
# Add these to your existing requirements.txt

# ==================== REST API ====================
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.3
django-cors-headers>=4.3.0

# ==================== Rate Limiting ====================
django-ratelimit>=4.1.0

# ==================== API Documentation (Optional) ====================
# Uncomment if you want OpenAPI/Swagger documentation
# drf-spectacular>=0.27.0

# ==================== Enhanced Features (Optional) ====================
# Uncomment for additional features

# Redis (for caching and Celery) - already in main requirements
# redis>=5.0.0

# Celery (for async tasks) - already in main requirements
# celery>=5.3.0

# Monitoring and Performance
# django-debug-toolbar>=4.2.0  # Development only
# django-silk>=5.0.0  # API profiling

# Advanced Security
# django-csp>=3.7  # Content Security Policy
# django-permissions-policy>=4.17.0

# Database Optimization
# django-cachalot>=2.6.0  # Query caching
# django-redis>=5.4.0  # Redis cache backend

# API Versioning
# django-rest-framework-versioning>=0.1.1

# Pagination
# django-rest-framework-link-header-pagination>=1.0.0

# File Upload handling
# django-chunked-upload>=2.0.0
# boto3>=1.28.0  # For S3 uploads

# Testing
# pytest-django>=4.5.0
# factory-boy>=3.3.0
# faker>=19.0.0

# Code Quality
# pylint>=3.0.0
# black>=23.0.0
# isort>=5.12.0

---

# Installation Instructions:

# 1. Backup current environment
pip freeze > requirements_old.txt

# 2. Install new dependencies
pip install -r requirements_additions.txt

# 3. Or install manually:
pip install djangorestframework djangorestframework-simplejwt django-filter django-cors-headers django-ratelimit

# 4. Verify installation
python -c "import rest_framework; print('DRF:', rest_framework.__version__)"
python -c "import rest_framework_simplejwt; print('JWT installed')"
python -c "import django_filters; print('Filters installed')"
python -c "import corsheaders; print('CORS installed')"
python -c "import ratelimit; print('Ratelimit installed')"

# 5. Run migrations
python manage.py makemigrations api notifications analytics
python manage.py migrate

# 6. Create superuser
python manage.py createsuperuser

# 7. Test API
python manage.py runserver
# Visit: http://localhost:8000/api/

---

# Docker Installation (if using Docker):

# Update Dockerfile:
# RUN pip install -r requirements_additions.txt

# Rebuild:
# docker-compose build
# docker-compose up
